Backend to develop

+++ backend/gym/src/main/java/c1220ftjavareact/gym/domain/dto/SubscriptionDTO.java
@@ -1,7 +1,9 @@
 package c1220ftjavareact.gym.domain.dto;
 
 import c1220ftjavareact.gym.repository.entity.State;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.time.LocalDateTime;
 
@@ -9,8 +11,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SubscriptionDTO {
-    private int idClient;
-    private int idClass;
+    private Long idClient;
+    private Long idClass;
     private State state;
     private LocalDateTime subscriptionDay;
 }
